Refactor exports and imports
