+++ src/components/SliderCard.tsx
@@ -1,5 +1,6 @@
-import styled from 'styled-components';
 import React from 'react';
+
+import styled from 'styled-components';
 
 export const HorizontalScroll = styled.div`
   display: flex;
