+++ src/components/index.ts
@@ -1,7 +1,16 @@
-export { default } from './Card';
+export { HomeCard } from './Card';
 export {
   HorizontalScroll,
   HorizontalSeparator,
   VerticalSeparator,
   SliderCard,
 } from './SliderCard';
+export { Footer } from './Footer';
+export { Carousel } from './Carousel';
+export { CarouselArrow } from './CarouselArrow';
+export { MapPin } from './MapPin';
+export { Navbar } from './Navbar';
+export { NavButton } from './NavButton';
+export { ServiceAspect } from './ServiceAspect';
+export { ServiceCard } from './ServiceCard';
+export { ServiceNavButton } from './ServiceNavButton';
